Move checkout and payment hooks
