+++ assets/js/base/components/cart-checkout/place-order-button/index.js
@@ -2,7 +2,7 @@
  * External dependencies
  */
 import { __ } from '@wordpress/i18n';
-import { useCheckoutSubmit } from '@woocommerce/base-hooks';
+import { useCheckoutSubmit } from '@woocommerce/base-context/hooks';
 import { Icon, done } from '@woocommerce/icons';
 import Button from '@woocommerce/base-components/button';
