+++ assets/js/base/context/cart-checkout/payment-methods/use-payment-method-registration.ts
@@ -7,8 +7,12 @@ import {
 	getExpressPaymentMethods,
 } from '@woocommerce/blocks-registry';
 import { useState, useEffect, useRef, useCallback } from '@wordpress/element';
-import { useEmitResponse, useShallowEqual } from '@woocommerce/base-hooks';
-import { useStoreCart, useStoreNotices } from '@woocommerce/base-context/hooks';
+import { useShallowEqual } from '@woocommerce/base-hooks';
+import {
+	useStoreCart,
+	useStoreNotices,
+	useEmitResponse,
+} from '@woocommerce/base-context/hooks';
 import {
 	CURRENT_USER_IS_ADMIN,
 	PAYMENT_GATEWAY_SORT_ORDER,
