+++ assets/js/base/context/hooks/index.js
@@ -1,6 +1,8 @@
 export * from './cart';
+export * from './checkout';
 export * from './collections';
 export * from './shipping';
+export * from './payment-methods';
 export * from './use-store-notices';
 export * from './use-store-events';
 export * from './use-query-state';
