+++ assets/js/base/hooks/index.js
@@ -1,5 +1,3 @@
-export * from './checkout';
-export * from './payment-methods';
 export * from './use-container-queries';
 export * from './use-local-storage-state';
 export * from './use-position-relative-to-viewport';
