+++ assets/js/blocks/active-filters/block.js
@@ -2,7 +2,7 @@
  * External dependencies
  */
 import { __ } from '@wordpress/i18n';
-import { useQueryStateByKey } from '@woocommerce/base-hooks';
+import { useQueryStateByKey } from '@woocommerce/base-context/hooks';
 import { useMemo } from '@wordpress/element';
 import classnames from 'classnames';
 import PropTypes from 'prop-types';
