+++ docs/extensibility/checkout-flow-and-events.md
@@ -4,21 +4,21 @@ This document gives an overview of the flow for the checkout in the WooCommerce 
 
 ## Table of Contents <!-- omit in toc -->
 
-- [General Concepts](#general-concepts)
-  - [Tracking flow through status](#tracking-flow-through-status)
-    - [`CheckoutProvider` Exposed Statuses](#checkoutprovider-exposed-statuses)
-      - [Special States:](#special-states)
-    - [`ShippingProvider` Exposed Statuses](#shippingprovider-exposed-statuses)
-    - [`PaymentMethodDataProvider` Exposed Statuses](#paymentmethoddataprovider-exposed-statuses)
-  - [Emitting Events](#emitting-events)
-    - [`onCheckoutBeforeProcessing`](#oncheckoutbeforeprocessing)
-    - [`onPaymentProcessing`](#onpaymentprocessing)
-    - [`onCheckoutAfterProcessingWithSuccess`](#oncheckoutafterprocessingwithsuccess)
-    - [`onCheckoutAfterProcessingWithError`](#oncheckoutafterprocessingwitherror)
-    - [`onShippingRateSuccess`](#onshippingratesuccess)
-    - [`onShippingRateFail`](#onshippingratefail)
-    - [`onShippingRateSelectSuccess`](#onshippingrateselectsuccess)
-    - [`onShippingRateSelectFail`](#onshippingrateselectfail)
+-   [General Concepts](#general-concepts)
+    -   [Tracking flow through status](#tracking-flow-through-status)
+        -   [`CheckoutProvider` Exposed Statuses](#checkoutprovider-exposed-statuses)
+            -   [Special States:](#special-states)
+        -   [`ShippingProvider` Exposed Statuses](#shippingprovider-exposed-statuses)
+        -   [`PaymentMethodDataProvider` Exposed Statuses](#paymentmethoddataprovider-exposed-statuses)
+    -   [Emitting Events](#emitting-events)
+        -   [`onCheckoutBeforeProcessing`](#oncheckoutbeforeprocessing)
+        -   [`onPaymentProcessing`](#onpaymentprocessing)
+        -   [`onCheckoutAfterProcessingWithSuccess`](#oncheckoutafterprocessingwithsuccess)
+        -   [`onCheckoutAfterProcessingWithError`](#oncheckoutafterprocessingwitherror)
+        -   [`onShippingRateSuccess`](#onshippingratesuccess)
+        -   [`onShippingRateFail`](#onshippingratefail)
+        -   [`onShippingRateSelectSuccess`](#onshippingrateselectsuccess)
+        -   [`onShippingRateSelectFail`](#onshippingrateselectfail)
 
 The architecture of the Checkout Block is derived from the following principles:
 
@@ -165,7 +165,7 @@ const MyComponent = ( { onCheckoutBeforeProcessing } ) => {
 This internal hook contains a collection of API interfaces for event emitter related usage. It can be used by simply adding this to a component:
 
 ```jsx
-import { useEmitResponse } from '@woocommerce/base-hooks';
+import { useEmitResponse } from '@woocommerce/base-context/hooks';
 
 const Component = () => {
 	const {
